STYLESYNC_FLASK/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.5
python-dotenv==1.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.4
